+++ apps/ehr/src/features/visits/shared/components/SelectFromFavoritesButton.tsx
